Ensure integer dtype in labels test

+++ src/napari/layers/labels/_tests/test_labels_multiscale.py
@@ -88,10 +88,10 @@ def test_3D_multiscale_labels_in_3D():
 def test_multiscale_labels_data_setter():
     """Setting .data on a multiscale Labels layer should not crash."""
     shapes = [(40, 20), (20, 10), (10, 5)]
-    data = [np.zeros(s) for s in shapes]
+    data = [np.zeros(s, dtype=np.int8) for s in shapes]
     layer = Labels(data, multiscale=True)
 
-    new_data = [np.ones(s) for s in shapes]
+    new_data = [np.ones(s, dtype=np.int8) for s in shapes]
     layer.data = new_data
 
     assert layer.multiscale is True
